Updated seed nodes in akka.conf. Added Initializer trait and reworked all *App classes to initializers instances.
