+++ src/main/resources/akka.conf
@@ -9,7 +9,10 @@ akka {
     seed-nodes = [
       "akka.tcp://ClusterSystem@127.0.0.1:2551",
       "akka.tcp://ClusterSystem@127.0.0.1:2552",
-      "akka.tcp://ClusterSystem@127.0.0.1:2553"]
+      "akka.tcp://ClusterSystem@127.0.0.1:2553",
+      "akka.tcp://ClusterSystem@127.0.0.1:2554",
+      "akka.tcp://ClusterSystem@127.0.0.1:2555",
+      "akka.tcp://ClusterSystem@127.0.0.1:2556"]
 
     auto-down-unreachable-after = 10s
     min-nr-of-members = 2
