+++ src/main/resources/logback.xml
@@ -12,7 +12,7 @@
     <timestamp key="bySecond" datePattern="yyyy-MM-dd-HH-mm-ss"/>
 
     <appender name="FILE" class="ch.qos.logback.core.FileAppender">
-        <file>video-recommender-${bySecond}.log</file>
+        <file>${outout-file}-${bySecond}.log</file>
 
         <encoder>
             <pattern>%date %level [%thread] %logger{10} [%file:%line] %msg%n</pattern>
