+++ src/main/scala/ztis/Initializer.scala
@@ -1,0 +1,8 @@
+package ztis
+
+import com.typesafe.scalalogging.slf4j.StrictLogging
+import ztis.util.RouterSupport
+
+trait Initializer extends StrictLogging with RouterSupport {
+  def initialize(): Unit
+}
